Refactor top nav bar
