+++ src/app/components/app.jsx
@@ -2,10 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 //components
-import TopNav from './topNav.jsx'
-import SideNav from './sideNav.jsx';
-import MainHeader from './mainHeader.jsx';
-import Posts from './posts.jsx';
+import TopNav from './topNav/topNav.jsx';
+import SideNav from './sideNav/sideNav.jsx';
+import MainView from './mainView/mainView.jsx';
 
 export default class App extends React.Component {
   constructor() {
@@ -14,51 +13,13 @@ export default class App extends React.Component {
   render() {
     return (
       <div>
-        <div className="navbar navbar-default navbar-static-top">
-          <div className="container">
-            <div className="navbar-header">
-              <a href="../" className="navbar-brand">Code Wiki</a>
-            </div>
-            <nav className="collapse navbar-collapse">
-              <ul className="nav navbar-nav">
-                <li><a href="#">Post</a></li>
-              </ul>
-              <form className="navbar-form navbar-left" role="search">
-                <div className="form-group">
-                  <input type="text" className="form-control" placeholder="Search" />
-                </div>
-                <button type="submit" className="btn btn-default">Submit</button>
-              </form>
-              <ul className="nav navbar-nav navbar-right">
-                <li><a href="#">Sign In</a></li>
-                <li><a href="#">Sign Up</a></li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-          <SideNav />
-          <div className="col-sm-10 main-container">
-            <div className="main">
-              <MainHeader />
-              <div className="divider-full"></div>
-              <Posts />
-            </div>
-          </div>
         <TopNav />
+        <SideNav />
+        <MainView />
       </div>
     );
   }
 }
-
-
-
-
-
-
-
-
-
-
