+++ src/app/components/mainView/post.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom'
 
-export default class SideNav extends React.Component {
+export default class Post extends React.Component {
   constructor() {
     super();
   }
