--- src/app/components/searchBox.jsx
@@ -1,15 +1,0 @@
-import React from 'react'
-
-export default class SearchBox extends React.Component {
-  constructor() {
-    super();
-  }
-
-  render() {
-    return (
-      <span className='container'>
-      Search Box
-      </span>
-    );
-  }
-}
