+++ src/app/components/sideNav/sideNav.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 
 export default class SideNav extends React.Component {
   constructor() {
